tests/load_tests/requirements.txt
locust==2.15.1
python-socketio[client]==5.11.1
aiohttp==3.9.3 